+++ flask_jsonapi/__init__.py
@@ -3,10 +3,10 @@ from .api import Api
 from .filters_schema import FilterField
 from .filters_schema import FilterSchema
 from .filters_schema import ListFilterField
-from .resource_repositories import BaseResourceRepository
-from .resource_repositories import ResourceRepositoryDetailView
-from .resource_repositories import ResourceRepositoryListView
-from .resource_repositories import ResourceRepositoryViewSet
+from .resource_repositories.repositories import ResourceRepository
+from .resource_repository_views import ResourceRepositoryDetailView
+from .resource_repository_views import ResourceRepositoryListView
+from .resource_repository_views import ResourceRepositoryViewSet
 from .resources import ResourceDetail
 from .resources import ResourceList
 
@@ -16,7 +16,7 @@ __all__ = [
     FilterField,
     FilterSchema,
     ListFilterField,
-    BaseResourceRepository,
+    ResourceRepository,
     ResourceRepositoryDetailView,
     ResourceRepositoryListView,
     ResourceRepositoryViewSet,